Create a `useProfile()` hook, to refresh the profile at any time
